requirements.txt
speech_recognition>=3.8
python-dotenv>=0.19
pyttsx3>=2.90
streamlit>=1.0
PyPDF2>=1.26
requests>=2.26
